Change gap buffer package
