+++ echo-core/src/commonMain/kotlin/io/github/alexandrepiveteau/echo/core/buffer/Gap.kt
@@ -1,4 +1,4 @@
-package io.github.alexandrepiveteau.echo.core.internal.buffer
+package io.github.alexandrepiveteau.echo.core.buffer
 
 /** An interface representing the [Gap] from a gap buffer. */
 interface Gap {
