+++ echo-core/src/commonMain/kotlin/io/github/alexandrepiveteau/echo/core/buffer/MutableByteGapBuffer+Extensions.kt
@@ -1,4 +1,4 @@
-package io.github.alexandrepiveteau.echo.core.internal.buffer
+package io.github.alexandrepiveteau.echo.core.buffer
 
 /** Creates an empty [MutableByteGapBuffer]. */
 fun mutableByteGapBufferOf(): MutableByteGapBuffer = MutableByteGapBufferImpl()
