+++ echo-core/src/commonTest/kotlin/io/github/alexandrepiveteau/echo/core/buffer/MutableByteGapBufferFuzzTest.kt
@@ -1,9 +1,6 @@
-package io.github.alexandrepiveteau.echo.core.internal
+package io.github.alexandrepiveteau.echo.core.buffer
 
 import io.github.alexandrepiveteau.echo.core.Performance
-import io.github.alexandrepiveteau.echo.core.internal.buffer.MutableByteGapBuffer
-import io.github.alexandrepiveteau.echo.core.internal.buffer.mutableByteGapBufferOf
-import io.github.alexandrepiveteau.echo.core.internal.buffer.toByteArray
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
